requirements.txt
langchain==1.0.0a14~
langchain-openai==1.0.0a4
python-dotenv==1.1.1
